The globalStandardSkin setting is placed on the meshRenderer component
