+++ cocos/rendering/pipeline-scene-data.ts
@@ -34,7 +34,7 @@ import { Pass } from '../render-scene/core/pass';
 import { CSMLayers } from './shadow/csm-layers';
 import { cclegacy } from '../core';
 import { Skin } from '../render-scene/scene/skin';
-import { ModelRenderer } from '../misc/model-renderer';
+import { MeshRenderer } from '../3d';
 
 const GEOMETRY_RENDERER_TECHNIQUE_COUNT = 6;
 
@@ -80,7 +80,7 @@ export class PipelineSceneData {
      * @zh 设置一个全局的4s标准模型
      * @returns The model id
      */
-    set standardSkinModel (val: ModelRenderer | null) {
+    set standardSkinModel (val: MeshRenderer | null) {
         if (this._standardSkinModel && this._standardSkinModel !== val) this._standardSkinModel.closedStandardSkin();
         this._standardSkinModel = val;
     }
@@ -118,7 +118,7 @@ export class PipelineSceneData {
     protected _isHDR = true;
     protected _shadingScale = 1.0;
     protected _csmSupported = true;
-    private _standardSkinModel: ModelRenderer | null = null;
+    private _standardSkinModel: MeshRenderer | null = null;
 
     constructor () {
         this._shadingScale = 1.0;
